services/rpa-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
selenium==4.16.0
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
asyncio
typing-extensions
pathlib
json5==0.9.14
websockets==12.0
httpx==0.25.2
structlog==23.2.0
colorlog==6.8.0
sqlalchemy==2.0.23
asyncpg==0.29.0
